remove remote parts
